third_party/WebKit/LayoutTests/imported/wpt/dom/nodes/Document-createElementNS-expected.txt

This is a testharness.js-based test.
PASS createElementNS test in HTML document: null,null,null 
FAIL createElementNS test in XML document: null,null,null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,null,null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,undefined,null 
FAIL createElementNS test in XML document: null,undefined,null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,undefined,null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"foo",null 
FAIL createElementNS test in XML document: null,"foo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"foo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"1foo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"f1oo",null 
FAIL createElementNS test in XML document: null,"f1oo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"f1oo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"foo1",null 
FAIL createElementNS test in XML document: null,"foo1",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"foo1",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"ெfoo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"ெfoo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"ெfoo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"}foo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"}foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"}foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"f}oo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"f}oo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"f}oo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"foo}","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"foo}","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"foo}","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"\ufffffoo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"\ufffffoo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"\ufffffoo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"f\uffffoo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"f\uffffoo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"f\uffffoo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"foo\uffff","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"foo\uffff","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"foo\uffff","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"<foo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"<foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"<foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"foo>","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"foo>","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"foo>","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"<foo>","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"<foo>","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"<foo>","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"f<oo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"f<oo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"f<oo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"^^","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"^^","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"^^","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"fo o","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"fo o","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"fo o","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"-foo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"-foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"-foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,".foo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,".foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,".foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,":foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: null,":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"f:oo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: null,"f:oo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"f:oo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"foo:","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: null,"foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"f:o:o","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: null,"f:o:o","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"f:o:o","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,":","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: null,":","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,":","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"xml",null 
FAIL createElementNS test in XML document: null,"xml",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"xml",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"xmlns","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: null,"xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"xmlfoo",null 
FAIL createElementNS test in XML document: null,"xmlfoo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"xmlfoo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"xml:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: null,"xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"xmlns:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: null,"xmlns:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"xmlns:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"xmlfoo:bar","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: null,"xmlfoo:bar","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"xmlfoo:bar","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"null:xml","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: null,"null:xml","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"null:xml","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "",null,null 
FAIL createElementNS test in XML document: "",null,null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "",null,null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "",":foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "",":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "",":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "","f:oo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "","f:oo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "","f:oo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "","foo:","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "","foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "","foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,null,null 
FAIL createElementNS test in XML document: undefined,null,null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,null,null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,undefined,null 
FAIL createElementNS test in XML document: undefined,undefined,null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,undefined,null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"foo",null 
FAIL createElementNS test in XML document: undefined,"foo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"foo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"1foo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: undefined,"1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"f1oo",null 
FAIL createElementNS test in XML document: undefined,"f1oo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"f1oo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"foo1",null 
FAIL createElementNS test in XML document: undefined,"foo1",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"foo1",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,":foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: undefined,":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"f:oo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: undefined,"f:oo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"f:oo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"foo:","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: undefined,"foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"f::oo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: undefined,"f::oo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"f::oo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"xml",null 
FAIL createElementNS test in XML document: undefined,"xml",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"xml",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"xmlns","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: undefined,"xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"xmlfoo",null 
FAIL createElementNS test in XML document: undefined,"xmlfoo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"xmlfoo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"xml:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: undefined,"xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"xmlns:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: undefined,"xmlns:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"xmlns:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"xmlfoo:bar","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: undefined,"xmlfoo:bar","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"xmlfoo:bar","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","foo",null 
FAIL createElementNS test in XML document: "http://example.com/","foo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","foo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","1foo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","<foo>","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","<foo>","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","<foo>","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","fo<o","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","fo<o","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","fo<o","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","-foo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","-foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","-foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/",".foo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/",".foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/",".foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","f1oo",null 
FAIL createElementNS test in XML document: "http://example.com/","f1oo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","f1oo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","foo1",null 
FAIL createElementNS test in XML document: "http://example.com/","foo1",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","foo1",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/",":foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://example.com/",":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/",":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","f:oo",null 
FAIL createElementNS test in XML document: "http://example.com/","f:oo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","f:oo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","f:o:o","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","f:o:o","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","f:o:o","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","foo:","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","f::oo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","f::oo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","f::oo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
FAIL createElementNS test in HTML document: "http://example.com/","a:0","NAMESPACE_ERR" assert_throws: function "function () { doc.createElementNS(namespace, qualifiedName) }" threw object "InvalidCharacterError: Failed to execute 'createElementNS' on 'Document': The qualified name provided ('a:0') contains the invalid name-start character '0'." that is not a DOMException NAMESPACE_ERR: property "code" is equal to 5, expected 14
FAIL createElementNS test in XML document: "http://example.com/","a:0","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","a:0","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","0:a","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","0:a","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","0:a","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","a:_",null 
FAIL createElementNS test in XML document: "http://example.com/","a:_",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","a:_",null assert_equals: expected "Dummy XHTML document" but got ""
FAIL createElementNS test in HTML document: "http://example.com/","a:ெ","NAMESPACE_ERR" assert_throws: function "function () { doc.createElementNS(namespace, qualifiedName) }" threw object "InvalidCharacterError: Failed to execute 'createElementNS' on 'Document': The qualified name provided ('a:ெ') contains the invalid name-start character 'ெ'." that is not a DOMException NAMESPACE_ERR: property "code" is equal to 5, expected 14
FAIL createElementNS test in XML document: "http://example.com/","a:ெ","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","a:ெ","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","ெ:a","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","ெ:a","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","ெ:a","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","a:aெ",null 
FAIL createElementNS test in XML document: "http://example.com/","a:aெ",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","a:aெ",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","aெ:a",null 
FAIL createElementNS test in XML document: "http://example.com/","aெ:a",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","aெ:a",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","xml:test","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","xml:test","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","xml:test","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","xmlns:test","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","xmlns:test","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","xmlns:test","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","test:xmlns",null 
FAIL createElementNS test in XML document: "http://example.com/","test:xmlns",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","test:xmlns",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","xmlns","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","_:_",null 
FAIL createElementNS test in XML document: "http://example.com/","_:_",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","_:_",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","_:h0",null 
FAIL createElementNS test in XML document: "http://example.com/","_:h0",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","_:h0",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","_:test",null 
FAIL createElementNS test in XML document: "http://example.com/","_:test",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","_:test",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","l_:_",null 
FAIL createElementNS test in XML document: "http://example.com/","l_:_",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","l_:_",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","ns:_0",null 
FAIL createElementNS test in XML document: "http://example.com/","ns:_0",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","ns:_0",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","ns:a0",null 
FAIL createElementNS test in XML document: "http://example.com/","ns:a0",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","ns:a0",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","ns0:test",null 
FAIL createElementNS test in XML document: "http://example.com/","ns0:test",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","ns0:test",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","a.b:c",null 
FAIL createElementNS test in XML document: "http://example.com/","a.b:c",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","a.b:c",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","a-b:c",null 
FAIL createElementNS test in XML document: "http://example.com/","a-b:c",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","a-b:c",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","xml",null 
FAIL createElementNS test in XML document: "http://example.com/","xml",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","xml",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","xmlns","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","XMLNS",null 
FAIL createElementNS test in XML document: "http://example.com/","XMLNS",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","XMLNS",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","xmlfoo",null 
FAIL createElementNS test in XML document: "http://example.com/","xmlfoo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","xmlfoo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","xml:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","XML:foo",null 
FAIL createElementNS test in XML document: "http://example.com/","XML:foo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","XML:foo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","xmlns:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","xmlns:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","xmlns:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","XMLNS:foo",null 
FAIL createElementNS test in XML document: "http://example.com/","XMLNS:foo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","XMLNS:foo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","xmlfoo:bar",null 
FAIL createElementNS test in XML document: "http://example.com/","xmlfoo:bar",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","xmlfoo:bar",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","prefix::local","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","prefix::local","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","prefix::local","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:{","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:{","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:{","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:}","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:}","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:}","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:~","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:~","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:~","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:'","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:'","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:'","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:!","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:!","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:!","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:@","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:@","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:@","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:#","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:#","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:#","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:$","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:$","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:$","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:%","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:%","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:%","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:^","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:^","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:^","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:&","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:&","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:&","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:*","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:*","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:*","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:(","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:(","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:(","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:)","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:)","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:)","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:+","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:+","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:+","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:=","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:=","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:=","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:[","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:[","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:[","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:]","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:]","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:]","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:\\","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:\\","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:\\","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:/","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:/","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:/","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:;","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:;","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:;","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:`","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:`","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:`","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:<","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:<","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:<","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:>","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:>","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:>","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:,","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:,","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:,","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:a ","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:a ","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:a ","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","namespaceURI:\"","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","namespaceURI:\"","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","namespaceURI:\"","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/","foo",null 
FAIL createElementNS test in XML document: "/","foo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/","foo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/","1foo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "/","1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/","1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/","f1oo",null 
FAIL createElementNS test in XML document: "/","f1oo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/","f1oo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/","foo1",null 
FAIL createElementNS test in XML document: "/","foo1",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/","foo1",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/",":foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "/",":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/",":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/","f:oo",null 
FAIL createElementNS test in XML document: "/","f:oo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/","f:oo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/","foo:","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "/","foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/","foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/","xml",null 
FAIL createElementNS test in XML document: "/","xml",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/","xml",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/","xmlns","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "/","xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/","xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/","xmlfoo",null 
FAIL createElementNS test in XML document: "/","xmlfoo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/","xmlfoo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/","xml:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "/","xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/","xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/","xmlns:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "/","xmlns:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/","xmlns:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/","xmlfoo:bar",null 
FAIL createElementNS test in XML document: "/","xmlfoo:bar",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/","xmlfoo:bar",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace","foo",null 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace","foo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace","foo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace","1foo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace","1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace","1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace","f1oo",null 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace","f1oo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace","f1oo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace","foo1",null 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace","foo1",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace","foo1",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace",":foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace",":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace",":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace","f:oo",null 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace","f:oo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace","f:oo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace","foo:","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace","foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace","foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace","xml",null 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace","xml",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace","xml",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace","xmlns","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace","xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace","xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace","xmlfoo",null 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace","xmlfoo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace","xmlfoo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace","xml:foo",null 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace","xml:foo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace","xml:foo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace","xmlns:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace","xmlns:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace","xmlns:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace","xmlfoo:bar",null 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace","xmlfoo:bar",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace","xmlfoo:bar",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespaces","xml:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespaces","xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespaces","xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/xml/1998/namespace","xml:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/xml/1998/namespace","xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/xml/1998/namespace","xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","1foo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","f1oo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","f1oo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","f1oo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","foo1","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","foo1","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","foo1","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/",":foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/",":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/",":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","f:oo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","f:oo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","f:oo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","foo:","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","xml","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","xml","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","xml","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","xmlns",null 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","xmlns",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","xmlns",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","xmlfoo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","xmlfoo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","xmlfoo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","xml:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","xmlns:foo",null 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","xmlns:foo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","xmlns:foo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","xmlfoo:bar","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","xmlfoo:bar","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","xmlfoo:bar","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","foo:xmlns","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","foo:xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","foo:xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:","foo",null 
FAIL createElementNS test in XML document: "foo:","foo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:","foo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:","1foo","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "foo:","1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:","1foo","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:","f1oo",null 
FAIL createElementNS test in XML document: "foo:","f1oo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:","f1oo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:","foo1",null 
FAIL createElementNS test in XML document: "foo:","foo1",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:","foo1",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:",":foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "foo:",":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:",":foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:","f:oo",null 
FAIL createElementNS test in XML document: "foo:","f:oo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:","f:oo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:","foo:","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "foo:","foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:","foo:","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:","xml",null 
FAIL createElementNS test in XML document: "foo:","xml",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:","xml",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:","xmlns","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "foo:","xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:","xmlns","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:","xmlfoo",null 
FAIL createElementNS test in XML document: "foo:","xmlfoo",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:","xmlfoo",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:","xml:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "foo:","xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:","xml:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:","xmlns:foo","NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "foo:","xmlns:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:","xmlns:foo","NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:","xmlfoo:bar",null 
FAIL createElementNS test in XML document: "foo:","xmlfoo:bar",null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:","xmlfoo:bar",null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "","","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "","","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "","","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,null,null 
FAIL createElementNS test in XML document: null,null,null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,null,null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: null,"","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: null,"","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: null,"","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,null,null 
FAIL createElementNS test in XML document: undefined,null,null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,null,null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: undefined,"","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: undefined,"","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: undefined,"","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/",null,null 
FAIL createElementNS test in XML document: "http://example.com/",null,null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/",null,null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://example.com/","","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://example.com/","","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://example.com/","","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/",null,null 
FAIL createElementNS test in XML document: "/",null,null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/",null,null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "/","","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "/","","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "/","","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace",null,null 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace",null,null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace",null,null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/XML/1998/namespace","","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/XML/1998/namespace","","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/XML/1998/namespace","","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/",null,"NAMESPACE_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/",null,"NAMESPACE_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/",null,"NAMESPACE_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "http://www.w3.org/2000/xmlns/","","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "http://www.w3.org/2000/xmlns/","","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "http://www.w3.org/2000/xmlns/","","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:",null,null 
FAIL createElementNS test in XML document: "foo:",null,null assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:",null,null assert_equals: expected "Dummy XHTML document" but got ""
PASS createElementNS test in HTML document: "foo:","","INVALID_CHARACTER_ERR" 
FAIL createElementNS test in XML document: "foo:","","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XML document" but got ""
FAIL createElementNS test in XHTML document: "foo:","","INVALID_CHARACTER_ERR" assert_equals: expected "Dummy XHTML document" but got ""
PASS Lower-case HTML element without a prefix 
PASS Lower-case HTML element with a prefix 
PASS Lower-case non-HTML element without a prefix 
PASS Lower-case non-HTML element with a prefix 
PASS Upper-case HTML element without a prefix 
PASS Upper-case HTML element with a prefix 
PASS Upper-case non-HTML element without a prefix 
PASS Upper-case non-HTML element with a prefix 
PASS null namespace 
PASS undefined namespace 
PASS empty string namespace 
Harness: the test ran to completion.
